Prepare a fix for #780.
