+++ postgresql/tox.ini
@@ -4,10 +4,10 @@ envlist = {py37,py310,pypy3}-{server,server680,server_postgresql_ssl,web},py37-d
 
 [testenv]
 commands =
-    server: py.test --instafail --slow --backend devpi_postgresql.main {posargs} --pyargs ../server/test_devpi_server
-    server680: py.test --instafail --slow --backend devpi_postgresql.main {posargs} --pyargs {envsitepackagesdir}/test_devpi_server
-    server_postgresql_ssl: py.test --instafail --backend devpi_postgresql.main --backend-postgresql-ssl  {posargs} --pyargs ../server/test_devpi_server/test_auth.py::TestAuth::test_auth_direct
-    web: py.test --instafail --backend devpi_postgresql.main {posargs} ../web/tests
+    server: py.test -p pytest_devpi_server -p pytest_devpi_postgresql --instafail --slow --backend devpi_postgresql.main {posargs} --pyargs ../server/test_devpi_server
+    server680: py.test -p pytest_devpi_server -p pytest_devpi_postgresql --instafail --slow --backend devpi_postgresql.main {posargs} --pyargs {envsitepackagesdir}/test_devpi_server
+    server_postgresql_ssl: py.test -p pytest_devpi_server -p pytest_devpi_postgresql --instafail --backend devpi_postgresql.main --backend-postgresql-ssl  {posargs} --pyargs ../server/test_devpi_server/test_auth.py::TestAuth::test_auth_direct
+    web: py.test -p pytest_devpi_server -p pytest_devpi_postgresql --instafail --backend devpi_postgresql.main {posargs} ../web/tests
 passenv = GITHUB_ACTIONS, LANG, USER
 setenv =
     dont_use_copy: DEVPI_PG_USE_COPY=false
