+++ server/.flake8
@@ -1,7 +1,7 @@
 [flake8]
 ignore = E501,E741,W503
 per-file-ignores =
-    conftest.py:E127,E128,E225,E231,E265,E402,E712
+    plugin.py:E127,E128,E225,E231,E265,E402,E712
     importexport.py:E127
     model.py:E128,E225,E231
     setup.py:E121,E126
